add support for triggering parameters
this allows setting the following AcquisitionControl parameters:
* TriggerMode
* TriggerSource
* LineSelector
* LineMode
* LineSource

with this it is possible to configure cameras to either issue trigger
commands or receive them (or both).

see e.g. also [this post] from VAimaging about how to set this up with
their cameras.

this was done as a quick-win, IMHO the proper solution is to implement
support for arbitrary parameters (see #57).

[this post]: https://va-imaging.com/blogs/machine-vision-knowledge-center/i-o-control-how-to-hardware-trigger-a-machine-vision-camera-using-a-trigger-sensor-and-how-to-trigger-machine-vision-lights